Allow users to edit more properties in the QR code and update everything live

+++ 2FAQRCodeDemo/Extensions.cs
@@ -8,7 +8,7 @@ namespace _2FAQRCodeDemo
 {
     public static class Extensions
     {
-        public static int LimitRange(this int i, int minInclusive, int maxInclusive)
+        public static int LimitRange(this ushort i, ushort minInclusive, ushort maxInclusive)
         {
             return Math.Max(Math.Min(i, maxInclusive), minInclusive);
         }
